[Refactor] rename to a more intuitive component name
